develope3: tests 14

+++ controllers/cards.js
@@ -1,23 +1,32 @@
+const mongoose = require('mongoose')
 const {
   Card,
 } = require('../models/card')
+
 const {
-  handleError,
-} = require('../utils/handleError')
+  NotFoundError,
+} = require('../errors/NotFoundError')
+// const {
+//  UnauthorizedError
+// } = require('../errors/UnauthorizedError')
+
+const {
+  ForbiddenError,
+} = require('../errors/ForbiddenError')
 
 // GET /cards — возвращает все карточки
-async function getAllCards(req, res) {
+async function getAllCards(req, res, next) {
   try {
     const cards = await Card.find({
     })
     res.send(cards)
   } catch (err) {
-    handleError(err, req, res)
+    next(err)
   }
 }
 
 // POST /cards — создаёт карточку
-async function createCard(req, res) {
+async function createCard(req, res, next) {
   try {
     const {
       name, link,
@@ -28,33 +37,44 @@ async function createCard(req, res) {
     })
     res.send(card)
   } catch (err) {
-    handleError(err, req, res)
+    next(err)
   }
 }
 
 // DELETE /cards/:cardId — удаляет карточку по идентификатору
-async function deleteCard(req, res) {
+async function deleteCard(req, res, next) {
   try {
     const {
       cardId,
     } = req.params
 
-    const card = await Card.findByIdAndRemove(cardId)
+    let card = mongoose.Types.ObjectId.isValid(cardId)
+
+    if (card) {
+      card = await Card.findById(cardId).populate('owner')
+    }
 
     if (!card) {
-      const error = new Error('Карточка не найдена')
-      error.name = 'NotFoundError'
-      throw error
+      throw new NotFoundError('Карточка не найдена')
     }
+
+    const ownerId = card.owner.id
+    const userId = req.user._id
+
+    if (ownerId !== userId) {
+      throw new ForbiddenError('Нельзя удалить чужую карточку')
+    }
+
+    await Card.findByIdAndRemove(cardId)
 
     res.send(card)
   } catch (err) {
-    handleError(err, req, res)
+    next(err)
   }
 }
 
 // PUT /cards/:cardId/likes — поставить лайк карточке
-async function putLike(req, res) {
+async function putLike(req, res, next) {
   try {
     const userId = req.user._id
     const card = await Card.findByIdAndUpdate(
@@ -70,18 +90,16 @@ async function putLike(req, res) {
     )
 
     if (!card) {
-      const error = new Error('Карточка не найдена')
-      error.name = 'NotFoundError'
-      throw error
+      throw new NotFoundError('Карточка не найдена')
     }
     res.send(card)
   } catch (err) {
-    handleError(err, req, res)
+    next(err)
   }
 }
 
 // DELETE /cards/:cardId/likes — убрать лайк с карточки
-async function deleteLike(req, res) {
+async function deleteLike(req, res, next) {
   try {
     const userId = req.user._id
     const card = await Card.findByIdAndUpdate(
@@ -96,13 +114,11 @@ async function deleteLike(req, res) {
       },
     )
     if (!card) {
-      const error = new Error('Карточка не найдена')
-      error.name = 'NotFoundError'
-      throw error
+      throw new NotFoundError('Карточка не найдена')
     }
     res.send(card)
   } catch (err) {
-    handleError(err, req, res)
+    next(err)
   }
 }
